+++ src/main/java/com/artipie/api/ssl/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * The MIT License (MIT) Copyright (c) 2020-2021 artipie.com
+ * https://github.com/artipie/artipie/LICENSE.txt
+ */
+
+/**
+ * Artipie Rest API.
+ *
+ * @since 0.26
+ */
+package com.artipie.api.ssl;
